backend/pyproject.toml
[project]
name = "macroai-backend"
version = "0.1.0"
description = "MacroAI API server"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "beanie>=1.26",
    "motor>=3.6",
    "redis[hiredis]>=5.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "python-multipart>=0.0.9",
    "langgraph>=1.0",
    "langchain-core>=0.3",
    "langchain-community>=0.3",
    "langgraph-checkpoint-mongodb>=0.3",
    "litellm>=1.81",
    "cryptography>=43.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.8",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
